Simplify signup so that services can immediately configure everything (#1584)

+++ src/shop/jobs/clean_unverified_signups.py
@@ -21,22 +21,30 @@ from google.appengine.ext import db
 
 from rogerthat.bizz.job import run_job
 from rogerthat.utils import now
-from shop.models import CustomerSignup
+from shop.models import CustomerSignup, CustomerSignupStatus
 
 
 def all_pending_signups():
+    # TODO: 1 week after first week of june 2020, this function can be replaced by all_pending_signups2
     # unverified signup is the one that its email address hasn't been verified yet
     # not done and doesn't have an inbox message, as inbox messages are sent on email successful verification
     return CustomerSignup.all(keys_only=True).filter('done', False).filter('inbox_message_key', None)
 
 
+def all_pending_signups2():
+    return CustomerSignup.all(keys_only=True).filter('status', CustomerSignupStatus.PENDING)
+
+
 def remove_if_expired(signup_key, at):
     signup = CustomerSignup.get(signup_key)
     timestamp = signup.timestamp
-    if not timestamp or (at - timestamp) > CustomerSignup.EXPIRE_TIME:
+    # Delete signups which have not verified their email after a certain time
+    # If they have verified their email, inbox_message_key will be set
+    if not signup.inbox_message_key and not timestamp or (at - timestamp) > CustomerSignup.EXPIRE_TIME:
         logging.info('Deleting CustomerSignup:\n%s', db.to_dict(signup))
         db.delete(signup)
 
 
 def job():
     run_job(all_pending_signups, [], remove_if_expired, [now()])
+    run_job(all_pending_signups2, [], remove_if_expired, [now()])
